fixed path registry ckpt
