+++ mdagent/utils/path_registry.py
@@ -22,8 +22,10 @@ class PathRegistry:
 
     @classmethod
     def get_instance(cls, ckpt_dir: str = "ckpt"):
+        # caveat: if instance already exists, "ckpt_dir" argument is ignored
         if not cls.instance:
             cls.instance = cls(ckpt_dir)
+        cls.instance.ckpt_dir = cls.set_ckpt.set_ckpt_subdir(ckpt_dir=ckpt_dir)
         return cls.instance
 
     def __init__(self, ckpt_dir: str = "ckpt"):
